Further reduce scheduler compile memory
