+++ src/scheduling.cpp
@@ -45,41 +45,31 @@
 #include "rp1_route_bridge.hpp"
 #include "WSPR-Transmitter/src/rp1_gpclk_development_policy.hpp"
 #include "WSPR-Transmitter/src/rp1_gpclk_planner.hpp"
-#include "test_tone_frequency_plan.hpp"
-#include "test_tone_selector_plan.hpp"
 
 // Project headers
-#include "arg_parser.hpp"
 #include "band_gpio_selector.hpp"
+#include "band_lookup.hpp"
 #include "runtime_config_bridge.hpp"
+#include "runtime_config_operations.hpp"
 #include "frequency_semantics.hpp"
 #include "gpio_band_policy.hpp"
-#include "gpio_input.hpp"
 #include "gpio_output.hpp"
 #include "logging.hpp"
 #include "ppm_manager.hpp"
-#include "privileged_network_runtime.hpp"
-#include "support_request_guard.hpp"
-#include "signal_handler.hpp"
 #include "system_clock_frequency_estimate.hpp"
 #include "execution_plan_compiler.hpp"
 #include "wspr_reference_adapter.hpp"
-#include "web_server.hpp"
-#include "web_socket.hpp"
-#include "wspr_transmit.hpp"
+#include "transmitter_runtime_bridge.hpp"
 #include "version.hpp"
-#include "machine_power_control.hpp"
 #include "non_wspr_request_builder.hpp"
 
 // Standard library headers
 #include <algorithm>
 #include <atomic>
 #include <cctype>
-#include <cerrno>
 #include <chrono>
 #include <cmath>
 #include <condition_variable>
-#include <cstring>
 #include <ctime>
 #include <exception>
 #include <fstream>
@@ -95,9 +85,6 @@
 #include <mutex>
 
 // System headers
-#include <string.h>
-#include <sys/resource.h>
-#include <unistd.h>
 
 /**
  * @brief Selects and controls the GPIO assigned to the active amateur band.
@@ -297,6 +284,11 @@ namespace
     }
 }
 
+std::string active_gpio_log_suffix()
+{
+    return get_active_gpio_suffix();
+}
+
 void refresh_frequency_estimate_for_config()
 {
     refresh_frequency_estimate();
@@ -327,9 +319,9 @@ void commit_band_gpio_snapshot_to_request(
 static void commit_execution_request(
     const wsprrypi::TransmissionRequest &controller_request,
     const TransmissionRequest &legacy_request);
-static void clear_committed_execution_request() noexcept;
+void clear_committed_execution_request() noexcept;
 static bool refresh_committed_band_gpio_selection() noexcept;
-static void assert_transmit_gpio_outputs(const char *context) noexcept;
+void assert_transmit_gpio_outputs(const char *context) noexcept;
 void deassert_transmit_gpio_outputs(
     const ArgParserConfig *selector_config,
     bool keep_selector_gpio_initialized,
@@ -457,14 +449,6 @@ std::atomic<bool> shutdown_flag{false};
  */
 ModeType lastMode;
 
-enum class TestToneRestorationOwner
-{
-    Unknown,
-    WsprScheduler,
-    DirectToneStartup,
-    ManagedIdleNonWspr,
-};
-
 TestToneRestorationOwner test_tone_restoration_owner =
     TestToneRestorationOwner::Unknown;
 
@@ -477,7 +461,6 @@ TestToneRestorationOwner test_tone_restoration_owner =
 std::atomic<bool> web_test_tone{false};
 std::atomic<bool> shutdown_after_current_transmission{false};
 std::atomic<bool> shutdown_after_wspr_plan{false};
-static std::atomic<bool> transmission_runtime_failed{false};
 static bool managed_reload_tx_inhibited = false;
 bool suppress_scheduler_execution_for_test = false;
 static TestToneCommitInvokerForTest test_tone_commit_invoker_for_test{};
@@ -493,7 +476,7 @@ std::uint64_t non_wspr_schedule_generation_for_test() noexcept
 
 static std::atomic<BandGPIOPrepareStatus> active_band_gpio_prepare_status{
     BandGPIOPrepareStatus::Inactive};
-static std::atomic<bool> suppress_cancelled_ws_event_for_user_stop{false};
+std::atomic<bool> suppress_cancelled_ws_event_for_user_stop{false};
 static std::vector<SelectorGPIOReservation> idle_selector_gpio_reservations{};
 bool selector_gpio_control_enabled = false;
 bool selector_gpio_drive_enabled = false;
@@ -566,7 +549,7 @@ static bool configured_selector_gpio_contains(
     return false;
 }
 
-static bool runtime_should_hold_selector_gpios_initialized(
+bool runtime_should_hold_selector_gpios_initialized(
     const ArgParserConfig &cfg) noexcept
 {
     return selector_gpio_control_enabled &&
@@ -773,7 +756,7 @@ static std::vector<BandGPIOConfig> collect_selector_gpio_shutdown_targets(
     return targets;
 }
 
-static void shutdown_all_configured_selector_gpios(
+void shutdown_all_configured_selector_gpios(
     const ArgParserConfig &cfg) noexcept
 {
     std::vector<BandGPIOConfig> targets =
@@ -1086,7 +1069,7 @@ static void mark_tx_led_active_state(bool active) noexcept
     tx_led_state_cv.notify_all();
 }
 
-static void set_tx_led_state(bool state, const char *context) noexcept
+void set_tx_led_state(bool state, const char *context) noexcept
 {
     if (!should_control_tx_led())
     {
@@ -1141,7 +1124,7 @@ static void set_amp_gpio_state(bool state, const char *context) noexcept
     }
 }
 
-static void assert_transmit_gpio_outputs(const char *context) noexcept
+void assert_transmit_gpio_outputs(const char *context) noexcept
 {
     std::lock_guard<std::mutex> lk(transmit_gpio_lifecycle_mtx);
 
@@ -1218,7 +1201,7 @@ static bool reconcile_transmit_gpio_after_transmitter_stop(const char *context) 
     return fallback_used;
 }
 
-static bool reconcile_tx_led_after_transmitter_stop(const char *context) noexcept
+bool reconcile_tx_led_after_transmitter_stop(const char *context) noexcept
 {
     if (!should_control_tx_led())
     {
@@ -1402,7 +1385,7 @@ void commit_execution_request(
             return;
         }
 
-        wsprTransmitter.configureExecution(
+        transmitter_configure_execution(
             controller_request,
             current_transmission_request);
         return;
@@ -1418,7 +1401,7 @@ void commit_execution_request(
             return;
         }
 
-        wsprTransmitter.configureExecution(current_transmission_request);
+        transmitter_configure_execution(current_transmission_request);
         return;
     }
 
@@ -1443,7 +1426,7 @@ void commit_execution_request(
         committed_execution_route_for_test_storage =
             CommittedExecutionRouteForTest::CONTROLLER_WSPR;
 
-        wsprTransmitter.configureExecution(
+        transmitter_configure_execution(
             controller_request,
             current_transmission_request);
         return;
@@ -1451,7 +1434,7 @@ void commit_execution_request(
 
     committed_execution_route_for_test_storage =
         CommittedExecutionRouteForTest::LEGACY;
-    wsprTransmitter.configureExecution(current_transmission_request);
+    transmitter_configure_execution(current_transmission_request);
 }
 
 static wsprrypi::TransmissionMode to_controller_mode(ModeType mode) noexcept
@@ -1521,18 +1504,18 @@ static void commit_execution_request(
         return;
     }
 
-    wsprTransmitter.configureExecution(controller_request, current_transmission_request);
+    transmitter_configure_execution(controller_request, current_transmission_request);
 }
 
 /**
  * @brief Clear scheduler-owned execution snapshots after a completed stop.
  *
  * The transmitter owns a separate execution snapshot.  It is cleared by
- * WsprTransmitter::clearExecutionStateAfterStop(); the scheduler must also
+ * transmitter_clear_execution_state_after_stop(); the scheduler must also
  * discard its committed request so a completed transient Test Tone cannot be
  * reported or reused as live work.
  */
-static void clear_committed_execution_request() noexcept
+void clear_committed_execution_request() noexcept
 {
     current_transmission_request = TransmissionRequest{};
     current_controller_request_for_test_storage.reset();
@@ -1737,7 +1720,7 @@ void reset_active_wspr_plan_state()
     active_wspr_plan_in_progress = false;
 }
 
-static bool active_wspr_plan_has_more_frames_after_current() noexcept
+bool active_wspr_plan_has_more_frames_after_current() noexcept
 {
     return active_wspr_plan_in_progress &&
            (active_wspr_frame_index + 1U) < active_wspr_plan.frameCount();
@@ -1793,10 +1776,10 @@ static wsprrypi::StartupQuiesceResult invoke_startup_quiesce()
 {
     if (startup_quiesce_invoker_for_test)
         return startup_quiesce_invoker_for_test();
-    return wsprTransmitter.quiesceForStartup();
+    return transmitter_quiesce_for_startup();
 }
 
-static bool run_startup_quiesce_gate(const ArgParserConfig &cfg)
+bool run_startup_quiesce_gate(const ArgParserConfig &cfg)
 {
     const wsprrypi::StartupQuiesceResult result = invoke_startup_quiesce();
     if (!result.ok)
@@ -1947,10 +1930,10 @@ void set_managed_reload_tx_inhibited(
 
 bool transmitter_reload_should_defer() noexcept
 {
-    const WsprTransmitter::State state = wsprTransmitter.getState();
+    const WsprTransmitState state = transmitter_state();
 
-    if (state == WsprTransmitter::State::TRANSMITTING ||
-        state == WsprTransmitter::State::RECOVERING)
+    if (state == WsprTransmitState::TRANSMITTING ||
+        state == WsprTransmitState::RECOVERING)
     {
         return true;
     }
@@ -1958,8 +1941,8 @@ bool transmitter_reload_should_defer() noexcept
     // Direct-tone modes start immediately and can still be in the launch
     // handoff while the controller remains ENABLED. Treat that window as
     // active for reload purposes so INI edits do not cancel the live run.
-    return state == WsprTransmitter::State::ENABLED &&
-           wsprTransmitter.activeExecutionIsTone();
+    return state == WsprTransmitState::ENABLED &&
+           transmitter_active_execution_is_tone();
 }
 
 std::string transmitter_reload_defer_debug_snapshot()
@@ -2019,20 +2002,20 @@ std::string transmitter_reload_defer_debug_snapshot()
     };
 
     std::ostringstream oss;
-    const WsprTransmitter::State state = wsprTransmitter.getState();
+    const WsprTransmitState state = transmitter_state();
     const bool state_transmitting =
-        state == WsprTransmitter::State::TRANSMITTING;
+        state == WsprTransmitState::TRANSMITTING;
     const bool state_recovering =
-        state == WsprTransmitter::State::RECOVERING;
+        state == WsprTransmitState::RECOVERING;
     const bool state_enabled =
-        state == WsprTransmitter::State::ENABLED;
+        state == WsprTransmitState::ENABLED;
     const bool active_execution_is_tone =
-        wsprTransmitter.activeExecutionIsTone();
+        transmitter_active_execution_is_tone();
     const bool defer =
         state_transmitting ||
         state_recovering ||
         (state_enabled && active_execution_is_tone);
-    const auto runtime_status = wsprTransmitter.runtimeExecutionStatusSnapshot();
+    const auto runtime_status = transmitter_runtime_status();
     const TransmissionRequest committed_request =
         current_transmission_request_for_test();
     const std::optional<wsprrypi::TransmissionRequest> controller_request =
@@ -2083,33 +2066,33 @@ std::string transmitter_reload_defer_debug_snapshot()
         oss << "NONE";
     }
 
-    oss << ", transmitter_snapshot={" << wsprTransmitter.reloadDeferDebugState() << "}";
+    oss << ", transmitter_snapshot={" << transmitter_reload_defer_debug_state() << "}";
     return oss.str();
 }
 
-static bool scheduler_managed_transmission_active_for_test_tone() noexcept
+bool scheduler_managed_transmission_active_for_test_tone() noexcept
 {
-    const WsprTransmitter::State state = wsprTransmitter.getState();
-    if (state != WsprTransmitter::State::TRANSMITTING &&
-        state != WsprTransmitter::State::RECOVERING)
+    const WsprTransmitState state = transmitter_state();
+    if (state != WsprTransmitState::TRANSMITTING &&
+        state != WsprTransmitState::RECOVERING)
     {
         return false;
     }
 
-    return !wsprTransmitter.activeExecutionIsTone();
+    return !transmitter_active_execution_is_tone();
 }
 
-static bool scheduler_managed_transmission_enabled_for_test_tone() noexcept
+bool scheduler_managed_transmission_enabled_for_test_tone() noexcept
 {
     return runtime_transmit_enabled(config);
 }
 
-static void finalize_transmission_stop_cleanup(
+void finalize_transmission_stop_cleanup(
     const ArgParserConfig *selector_config,
     bool keep_selector_gpio_initialized,
     const char *led_reason,
-    bool clear_scheduler_latches = false,
-    bool emit_debug_log = false)
+    bool clear_scheduler_latches,
+    bool emit_debug_log)
 {
     if (emit_debug_log)
     {
@@ -2152,7 +2135,7 @@ WsprFrequencyEntry next_frequency_entry_from(
     return entry;
 }
 
-static WsprFrequencyEntry next_frequency_entry(bool reset);
+WsprFrequencyEntry next_frequency_entry(bool reset);
 
 /**
  * @brief Return the prepared plan for a single frame from a saved plan.
@@ -2428,7 +2411,7 @@ double maybe_apply_wspr_random_offset(
  * This request is fully committed at the orchestration layer. The
  * transmitter must not infer any additional policy from tone mode.
  */
-static TransmissionRequest make_tone_request(
+TransmissionRequest make_tone_request(
     const ArgParserConfig &cfg,
     double committed_ppm,
     double actual_rf_frequency_hz,
@@ -2456,7 +2439,7 @@ static TransmissionRequest make_tone_request(
  * request is committed or startAsync() can be reached.  Callers therefore
  * leave both RF and selector outputs inactive when preparation fails.
  */
-static bool start_direct_tone_execution(
+bool start_direct_tone_execution(
     const ArgParserConfig &cfg,
     const WsprFrequencyEntry &entry,
     double actual_rf_frequency_hz,
@@ -2556,7 +2539,7 @@ static bool start_direct_tone_execution(
     }
     else if (!suppress_scheduler_execution_for_test)
     {
-        wsprTransmitter.startAsync();
+        transmitter_start_async();
     }
 
     return true;
@@ -2612,7 +2595,7 @@ static bool prepare_and_commit_non_wspr_request(
     return true;
 }
 
-static bool start_non_wspr_transmission_now(const ArgParserConfig &cfg)
+bool start_non_wspr_transmission_now(const ArgParserConfig &cfg)
 {
     const double committed_ppm = cfg.ppm;
     std::string policy_error;
@@ -2666,7 +2649,7 @@ static bool start_non_wspr_transmission_now(const ArgParserConfig &cfg)
         }
         if (!suppress_scheduler_execution_for_test)
         {
-            wsprTransmitter.startAsync();
+            transmitter_start_async();
         }
         llog.logS(DEBUG, "Transmitting QRSS message.");
 
@@ -2680,7 +2663,7 @@ static bool start_non_wspr_transmission_now(const ArgParserConfig &cfg)
 
         llog.logS(DEBUG,
                   "- Frequency (MHz): ",
-                  wsprTransmitter.formatFrequencyMHz(frequency_hz));
+                  transmitter_format_frequency_mhz(frequency_hz));
 
         llog.logS(DEBUG,
                   "- Dot length (s): ",
@@ -2738,7 +2721,7 @@ static bool start_non_wspr_transmission_now(const ArgParserConfig &cfg)
         }
         if (!suppress_scheduler_execution_for_test)
         {
-            wsprTransmitter.startAsync();
+            transmitter_start_async();
         }
         llog.logS(DEBUG, "Transmitting FSKCW message.");
 
@@ -2752,7 +2735,7 @@ static bool start_non_wspr_transmission_now(const ArgParserConfig &cfg)
 
         llog.logS(DEBUG,
                   "- Mark frequency (MHz): ",
-                  wsprTransmitter.formatFrequencyMHz(mark_frequency_hz));
+                  transmitter_format_frequency_mhz(mark_frequency_hz));
 
         llog.logS(DEBUG,
                   "- Space frequency (Hz): ",
@@ -2760,7 +2743,7 @@ static bool start_non_wspr_transmission_now(const ArgParserConfig &cfg)
 
         llog.logS(DEBUG,
                   "- Space frequency (MHz): ",
-                  wsprTransmitter.formatFrequencyMHz(space_frequency_hz));
+                  transmitter_format_frequency_mhz(space_frequency_hz));
 
         llog.logS(DEBUG,
                   "- Dot length (s): ",
@@ -2818,7 +2801,7 @@ static bool start_non_wspr_transmission_now(const ArgParserConfig &cfg)
         }
         if (!suppress_scheduler_execution_for_test)
         {
-            wsprTransmitter.startAsync();
+            transmitter_start_async();
         }
         llog.logS(DEBUG, "Transmitting DFCW message.");
 
@@ -2832,7 +2815,7 @@ static bool start_non_wspr_transmission_now(const ArgParserConfig &cfg)
 
         llog.logS(DEBUG,
                   "- Dot frequency (MHz): ",
-                  wsprTransmitter.formatFrequencyMHz(dot_frequency_hz));
+                  transmitter_format_frequency_mhz(dot_frequency_hz));
 
         llog.logS(DEBUG,
                   "- Dash frequency (Hz): ",
@@ -2840,7 +2823,7 @@ static bool start_non_wspr_transmission_now(const ArgParserConfig &cfg)
 
         llog.logS(DEBUG,
                   "- Dash frequency (MHz): ",
-                  wsprTransmitter.formatFrequencyMHz(dash_frequency_hz));
+                  transmitter_format_frequency_mhz(dash_frequency_hz));
 
         llog.logS(DEBUG,
                   "- Dot length (s): ",
@@ -2995,37 +2978,6 @@ void commit_band_gpio_snapshot_to_request(
         resolution.config.enabled && resolution.config.gpio >= 0;
     request.selector_band = resolution.band;
     request.selector_gpio_config = resolution.config;
-}
-
-static std::string format_elapsed(double elapsed)
-{
-    if (elapsed == 0.0)
-        return std::string();
-
-    std::ostringstream oss;
-    oss << std::fixed
-        << std::setprecision(6)
-        << elapsed;
-    return oss.str();
-}
-
-constexpr LogLevel to_log_level(WsprTransmitter::LogLevel level)
-{
-    switch (level)
-    {
-    case WsprTransmitter::LogLevel::DEBUG:
-        return LogLevel::DEBUG;
-    case WsprTransmitter::LogLevel::INFO:
-        return LogLevel::INFO;
-    case WsprTransmitter::LogLevel::WARN:
-        return LogLevel::WARN;
-    case WsprTransmitter::LogLevel::ERROR:
-        return LogLevel::ERROR;
-    case WsprTransmitter::LogLevel::FATAL:
-        return LogLevel::FATAL;
-    }
-
-    return LogLevel::INFO; // Safe fallback
 }
 
 /**
@@ -3272,1499 +3224,6 @@ bool request_wspr_shutdown(std::string_view reason)
     return !already_requested;
 }
 
-void transmitter_cb(WsprTransmitter::TransmissionCallbackEvent event,
-                    WsprTransmitter::LogLevel level,
-                    const std::string &msg,
-                    double value)
-{
-    switch (event)
-    {
-    case WsprTransmitter::TransmissionCallbackEvent::STARTING:
-    {
-        const double frequency = value;
-
-        if (config.mode == ModeType::WSPR &&
-            (!active_wspr_plan_in_progress || active_wspr_frame_index == 0U))
-        {
-            consume_tx_iteration_if_needed();
-        }
-
-        assert_transmit_gpio_outputs("transmission start");
-
-        // Notify clients of start.
-        send_ws_message("transmit", "starting");
-
-        // Log messages.
-        if (!msg.empty() && frequency != 0.0)
-        {
-            llog.logS(to_log_level(level),
-                      "Started transmission (",
-                      msg,
-                      ") ",
-                      wsprTransmitter.formatFrequencyMHz(frequency),
-                      " MHz",
-                      get_active_gpio_suffix(),
-                      ".");
-        }
-        else if (frequency != 0.0)
-        {
-            if (config.mode == ModeType::QRSS)
-            {
-                llog.logS(to_log_level(level),
-                          "Started QRSS transmission at frequency: ",
-                          wsprTransmitter.formatFrequencyMHz(config.qrss.frequency_hz),
-                          " MHz",
-                          get_active_gpio_suffix(),
-                          ".");
-            }
-            else if (config.mode == ModeType::FSKCW)
-            {
-                llog.logS(to_log_level(level),
-                          "Started FSKCW transmission at mark frequency: ",
-                          wsprTransmitter.formatFrequencyMHz(config.fskcw.mark_frequency_hz),
-                          " MHz",
-                          get_active_gpio_suffix(),
-                          ".");
-            }
-            else if (config.mode == ModeType::DFCW)
-            {
-                llog.logS(to_log_level(level),
-                          "Started DFCW transmission at dot frequency: ",
-                          wsprTransmitter.formatFrequencyMHz(config.dfcw.dot_frequency_hz),
-                          " MHz",
-                          get_active_gpio_suffix(),
-                          ".");
-            }
-            else
-            {
-                llog.logS(to_log_level(level),
-                          "Started transmission: ",
-                          wsprTransmitter.formatFrequencyMHz(frequency),
-                          " MHz",
-                          get_active_gpio_suffix(),
-                          ".");
-            }
-        }
-        else if (!msg.empty())
-        {
-            llog.logS(to_log_level(level),
-                      "Started transmission (",
-                      msg,
-                      ").");
-        }
-        else
-        {
-            llog.logS(to_log_level(level),
-                      "Started transmission.");
-        }
-        break;
-    }
-
-    case WsprTransmitter::TransmissionCallbackEvent::PROGRESS:
-    {
-        send_ws_message(
-            "transmit",
-            "progress",
-            std::string(),
-            static_cast<int>(value));
-        break;
-    }
-
-    case WsprTransmitter::TransmissionCallbackEvent::COMPLETE:
-    {
-        const double elapsed = value;
-        bool do_config = true;
-        const bool deferred_reload_pending =
-            ini_reload_pending.load(std::memory_order_acquire);
-
-        const std::string s_elapsed = format_elapsed(elapsed);
-        if (!msg.empty() && elapsed != 0.0)
-        {
-            llog.logS(to_log_level(level),
-                      "Completed transmission (",
-                      msg,
-                      ") ",
-                      s_elapsed,
-                      " seconds.");
-        }
-        else if (elapsed != 0.0)
-        {
-            llog.logS(to_log_level(level),
-                      "Completed transmission: ",
-                      s_elapsed,
-                      " seconds.");
-        }
-        else if (!msg.empty())
-        {
-            llog.logS(to_log_level(level),
-                      "Completed transmission (",
-                      msg,
-                      ").");
-        }
-        else
-        {
-            llog.logS(to_log_level(level),
-                      "Completed transmission.");
-        }
-
-        finalize_transmission_stop_cleanup(
-            &config,
-            runtime_should_hold_selector_gpios_initialized(config),
-            "transmission completion");
-
-        // Notify the websocket clients.
-        send_ws_message("transmit", "finished");
-
-        const bool shutdown_when_idle =
-            shutdown_after_current_transmission.exchange(false, std::memory_order_acq_rel);
-        const bool shutdown_when_plan_finishes =
-            shutdown_after_wspr_plan.load(std::memory_order_acquire);
-
-        if (deferred_reload_pending)
-        {
-            reset_active_wspr_plan_state();
-        }
-        else if (do_config && active_wspr_plan_has_more_frames_after_current())
-        {
-            ++active_wspr_frame_index;
-        }
-        else if (active_wspr_plan_in_progress)
-        {
-            reset_active_wspr_plan_state();
-        }
-
-        if (shutdown_when_idle && do_config)
-        {
-            request_wspr_shutdown("completed configured TX iterations");
-            do_config = false;
-        }
-        else if (shutdown_when_plan_finishes && do_config &&
-                 !active_wspr_plan_in_progress)
-        {
-            shutdown_after_wspr_plan.store(false, std::memory_order_release);
-            request_wspr_shutdown("completed configured TX iterations");
-            do_config = false;
-        }
-        else if (deferred_reload_pending && do_config)
-        {
-            set_config();
-            do_config = false;
-        }
-        else if (do_config &&
-                 config.mode != ModeType::WSPR &&
-                 config.mode != ModeType::TONE &&
-                 !has_non_wspr_cli_startup_request(config.mode))
-        {
-            schedule_next_non_wspr_launch(config);
-            do_config = false;
-        }
-
-        // Set config will determine if we have work to do.
-        if (do_config)
-        {
-            set_config();
-        }
-
-        break;
-    }
-
-    case WsprTransmitter::TransmissionCallbackEvent::FAILED:
-    {
-        transmission_runtime_failed.store(true, std::memory_order_release);
-        const std::string reason = msg.empty()
-            ? "Transmission backend execution failed."
-            : msg;
-        llog.logS(ERROR, "Transmission failed: ", reason);
-
-        finalize_transmission_stop_cleanup(
-            &config,
-            false,
-            "transmission failure",
-            true,
-            true);
-        send_ws_message("transmit", "failed", reason);
-
-        if (config.use_ini)
-        {
-            set_managed_reload_tx_inhibited(true);
-            llog.logS(
-                ERROR,
-                "Transmit is inhibited until configuration is reloaded or the service is restarted.");
-        }
-        else
-        {
-            request_wspr_shutdown("transmission backend failure");
-        }
-        break;
-    }
-
-    case WsprTransmitter::TransmissionCallbackEvent::CANCELLED:
-    {
-        const double elapsed = value;
-        const std::string s_elapsed = format_elapsed(elapsed);
-        const bool suppress_ws_event =
-            suppress_cancelled_ws_event_for_user_stop.exchange(
-                false,
-                std::memory_order_acq_rel);
-
-        llog.logS(to_log_level(level),
-                  "Transmission canceled after ",
-                  s_elapsed,
-                  " seconds.");
-
-        finalize_transmission_stop_cleanup(
-            &config,
-            runtime_should_hold_selector_gpios_initialized(config),
-            "transmission cancellation",
-            true);
-        if (suppress_ws_event)
-        {
-            llog.logS(
-                DEBUG,
-                "Suppressing websocket canceled event because an explicit user stop will publish stopped.");
-        }
-        else
-        {
-            send_ws_message("transmit", "canceled");
-        }
-
-        break;
-    }
-
-    case WsprTransmitter::TransmissionCallbackEvent::SKIPPED:
-    {
-        if (!current_transmission_request.isSkipWindow())
-        {
-            llog.logS(
-                WARN,
-                "Ignoring unexpected SKIPPED transmitter callback for non-skip request.");
-            break;
-        }
-
-        // Return transmit GPIO outputs to the same idle path used by every
-        // terminal transmitter event.
-        deassert_transmit_gpio_outputs(
-            &config,
-            runtime_should_hold_selector_gpios_initialized(config),
-            "transmission skip");
-
-        if (!msg.empty())
-            llog.logS(to_log_level(level), msg, ".");
-        else
-            llog.logS(to_log_level(level), "Skipping transmission.");
-
-        // Notify websocket clients.
-        send_ws_message("transmit", "skipped");
-
-        shutdown_after_current_transmission.store(false, std::memory_order_release);
-        shutdown_after_wspr_plan.store(false, std::memory_order_release);
-        reset_active_wspr_plan_state();
-
-        // Advance to the next configured slot.
-        set_config();
-
-        break;
-    }
-
-    case WsprTransmitter::TransmissionCallbackEvent::LOGGING:
-    default:
-    {
-        if (!msg.empty())
-            llog.logS(to_log_level(level), msg);
-
-        break;
-    }
-    }
-}
-
-/**
- * @brief  Callback invoked when PPMManager has a new PPM reading.
- *
- * @details
- * Sets the `ppm_reload_pending` flag so that downstream consumers
- * will pick up the new value and marks the provider estimate as available
- * after the adapter has delivered a measurement.
- *
- * @param new_ppm  The latest PPM correction value (ignored here; reload
- *                 logic will pull it from PPMManager when needed).
- */
-void ppm_callback(double /*new_ppm*/)
-{
-    refresh_frequency_estimate();
-    // Notify other subsystems to reload/recalibrate with the fresh PPM.
-    ppm_reload_pending.store(true, std::memory_order_relaxed);
-
-    // Record that the provider adapter has produced a value.
-    if (!config.frequency_estimate_good)
-    {
-        llog.logS(DEBUG, "The system-clock frequency estimate provider has updated its value.");
-        config.frequency_estimate_good = true;
-    }
-}
-
-/**
- * @brief   Initialize the PPM subsystem.
- *
- * Registers the PPM callback, initializes the PPMManager, and handles
- * any returned status. Qualification remains a soft gate; unavailable or
- * converging provider state is handled by deterministic correction fallback.
- *
- * @return  true if initialization is considered successful;
- *          false only on a fatal PPM error (e.g. excessive drift).
- */
-bool ppm_init()
-{
-    bool retval = true;
-
-    // Register the PPM update callback
-    ppmManager.setPPMCallback(ppm_callback);
-
-    // Perform the normal initialization
-    PPMStatus status = ppmManager.initialize();
-
-    switch (status)
-    {
-    case PPMStatus::SUCCESS:
-        llog.logS(INFO, "PPM Manager initialized successfully.");
-        break;
-
-    case PPMStatus::WARNING_HIGH_PPM:
-        llog.logE(ERROR, "Measured PPM exceeds safe threshold.");
-        return false;
-
-    case PPMStatus::ERROR_CHRONY_NOT_FOUND:
-        llog.logE(WARN,
-                  "chrony estimate unavailable; GPIO correction fallback policy will apply.");
-        break;
-
-    case PPMStatus::ERROR_UNSYNCHRONIZED_TIME:
-        llog.logE(WARN, "System-clock frequency estimate provider is not synchronized; fallback policy will apply.");
-        break;
-
-    default:
-        llog.logE(WARN, "Unknown PPMStatus returned from initialize().");
-        break;
-    }
-
-    return retval;
-}
-
-/**
- * @brief Callback function triggered to perform a system shutdown sequence.
- *
- * @details
- * This function is intended to be called when a shutdown GPIO event is triggered.
- * It logs the event and calls shutdown_system().
- */
-void callback_shutdown_system()
-{
-    llog.logS(INFO, "Shutdown called by GPIO", config.shutdown_pin);
-    shutdown_system();
-}
-
-/**
- * @brief Perform a system shutdown sequence.
- *
- * @details
- * This function is intended to be called when a shutdown event is triggered.
- * It performs a visual blink pattern on the LED pin if configured, sets the
- * shutdown flags, and notifies all threads waiting on the shutdown condition
- * variable.
- *
- * Specifically:
- * - Toggles the LED 3 times with 200ms intervals.
- * - Sets `exitwspr_cv` to break out of the main transmission loop.
- * - Sets `shutdown_flag` to mark that a full system shutdown is in progress.
- *
- * @note
- * The LED toggling uses `ledControl.toggleGPIO()` and assumes the hardware
- * supports it.
- */
-void shutdown_system()
-{
-    shutdown_flag.store(true, std::memory_order_relaxed);
-    request_wspr_shutdown("system power-off requested");
-
-    if (config.use_led)
-    {
-        // Flash LED three times if configured
-        for (int i = 0; i < 3; ++i)
-        {
-            set_tx_led_state(true, "shutdown blink on");
-            std::this_thread::sleep_for(std::chrono::milliseconds(200));
-
-            set_tx_led_state(false, "shutdown blink off");
-            if (i < 2)
-            {
-                std::this_thread::sleep_for(std::chrono::milliseconds(200));
-            }
-        }
-    }
-}
-
-/**
- * @brief Perform a system reboot sequence.
- *
- * @details
- * This function is intended to be called when a reboot event is triggered.
- * It performs a visual blink pattern on the LED pin if configured, sets the
- * reboot flags, and notifies all threads waiting on the reboot condition
- * variable.
- *
- * Specifically:
- * - Toggles the LED 2 times with 100ms intervals.
- * - Sets `exitwspr_cv` to break out of the main transmission loop.
- * - Sets `reboot_flag` to mark that a full system reboot is in progress.
- *
- * @note
- * The LED toggling uses `ledControl.toggleGPIO()` and assumes the hardware supports it.
- */
-void reboot_system()
-{
-    reboot_flag.store(true, std::memory_order_relaxed);
-    request_wspr_shutdown("System reboot requested");
-
-    if (config.use_led)
-    {
-        // Flash LED two times if configured
-        for (int i = 0; i < 2; ++i)
-        {
-            set_tx_led_state(true, "reboot blink on");
-            std::this_thread::sleep_for(std::chrono::milliseconds(100));
-
-            set_tx_led_state(false, "reboot blink off");
-            if (i < 2)
-            {
-                std::this_thread::sleep_for(std::chrono::milliseconds(100));
-            }
-        }
-    }
-}
-
-/**
- * @brief Start a transient runtime tone using scheduler-owned setup.
- *
- * The scheduler stops any active run, reuses the first configured
- * frequency entry, prepares band-selector GPIO state, commits a tone request,
- * and starts the transmitter. When provided, the override is the final RF
- * frequency and does not receive WSPR dial-frequency offset. Tone mode here is
- * runtime-only behavior.
- */
-static void rollback_failed_test_tone_start(ModeType previous_mode) noexcept
-{
-    wsprrypi::invalidateRp1GpclkDevelopmentOperation();
-    try
-    {
-        wsprTransmitter.stopAndJoin();
-        wsprTransmitter.clearExecutionStateAfterStop();
-        finalize_transmission_stop_cleanup(
-            &config,
-            false,
-            "test tone start rejection",
-            true,
-            true);
-    }
-    catch (const std::exception &error)
-    {
-        llog.logE(
-            ERROR,
-            "Test tone rejection cleanup failed: " +
-                std::string(error.what()));
-    }
-    catch (...)
-    {
-        llog.logE(ERROR, "Test tone rejection cleanup failed.");
-    }
-
-    (void)reconcile_tx_led_after_transmitter_stop(
-        "test tone start rejection");
-    current_transmission_request = TransmissionRequest{};
-    current_controller_request_for_test_storage.reset();
-    committed_execution_route_for_test_storage =
-        CommittedExecutionRouteForTest::NONE;
-    current_dial_frequency = 0.0;
-    current_frequency_entry = WsprFrequencyEntry{};
-    web_test_tone.store(false);
-    config.mode = previous_mode;
-    test_tone_restoration_owner = TestToneRestorationOwner::Unknown;
-}
-
-TestToneStartResult start_test_tone(const TestToneRequest &tone_request)
-{
-    TestToneStartResult result;
-    result.source = tone_request.source;
-
-    if (web_test_tone.load())
-    {
-        result.already_active = true;
-        result.message = "Test tone is already active.";
-        return result;
-    }
-
-    if (scheduler_managed_transmission_active_for_test_tone())
-    {
-        result.blocked_by_active_transmission = true;
-        result.message =
-            "Stop transmissions before starting a test tone. Disable transmissions after the active transmission stops.";
-        llog.logS(WARN, result.message);
-        return result;
-    }
-
-    if (scheduler_managed_transmission_enabled_for_test_tone())
-    {
-        result.blocked_by_enabled_transmission = true;
-        result.message =
-            "Disable transmissions before starting a test tone.";
-        llog.logS(WARN, result.message);
-        return result;
-    }
-
-    // Explicit requests are completely planned from one accepted configuration
-    // snapshot before any runtime mutation.
-    const bool has_explicit_frequency_source =
-        tone_request.source == TestToneFrequencySource::WsprBand ||
-        tone_request.source == TestToneFrequencySource::CustomRf;
-    std::optional<TestTonePlanningConfigSnapshot> explicit_planning_snapshot;
-    std::optional<TestToneFrequencyPlan> explicit_frequency_plan;
-    std::optional<TestToneSelectorPlan> explicit_selector_plan;
-    if (has_explicit_frequency_source)
-    {
-        const TestTonePlanningConfigSnapshot planning_snapshot =
-            current_test_tone_planning_config_snapshot();
-        const auto frequency_plan = plan_explicit_test_tone_frequency(
-            tone_request,
-            planning_snapshot.wspr_audio_offset_hz,
-            planning_snapshot.wspr_frequency_profile,
-            planning_snapshot.wspr_band_preferences);
-        if (!frequency_plan)
-        {
-            result.message = frequency_plan.error;
-            return result;
-        }
-        const auto resolved_band = lookup.lookup_ham_band(
-            static_cast<double>(frequency_plan.plan->actual_rf_frequency_hz));
-        if (!resolved_band.has_value())
-        {
-            result.message = "Unable to resolve the planned RF band.";
-            return result;
-        }
-        const auto selector_plan = plan_test_tone_selector(
-            *resolved_band, planning_snapshot.wspr_frequency_entries,
-            planning_snapshot.band_gpio);
-        if (!selector_plan)
-        {
-            result.message = selector_plan.error;
-            return result;
-        }
-        explicit_planning_snapshot = planning_snapshot;
-        explicit_frequency_plan = *frequency_plan.plan;
-        explicit_selector_plan = selector_plan.plan;
-
-        const auto gpio_policy = wsprrypi::evaluate_gpio_band_policy(
-            to_controller_backend(planning_snapshot.transmit_backend),
-            static_cast<double>(frequency_plan.plan->actual_rf_frequency_hz),
-            wsprrypi::TransmissionMode::TONE,
-            planning_snapshot.allow_unqualified_frequency,
-            planning_snapshot.allow_non_amateur_frequency,
-            to_controller_profile(planning_snapshot.transmit_backend));
-        if (!gpio_policy.allowed)
-        {
-            result.message = gpio_policy.error;
-            llog.logS(WARN, result.message);
-            return result;
-        }
-    }
-
-    // Capture restoration ownership before Test Tone changes config.mode.  A
-    // managed, transmit-disabled non-WSPR mode is idle configuration, not a
-    // transient direct-tone startup request.
-    TestToneRestorationOwner restoration_owner = TestToneRestorationOwner::Unknown;
-    if (config.mode == ModeType::WSPR)
-    {
-        restoration_owner = TestToneRestorationOwner::WsprScheduler;
-    }
-    else
-    {
-        WsprFrequencyEntry startup_entry;
-        double startup_rf_frequency_hz = 0.0;
-        if (try_get_direct_tone_startup_request(startup_entry, startup_rf_frequency_hz))
-        {
-            restoration_owner = TestToneRestorationOwner::DirectToneStartup;
-        }
-        else if (config.use_ini && !runtime_transmit_enabled(config) &&
-                 (config.mode == ModeType::FSKCW || config.mode == ModeType::QRSS ||
-                  config.mode == ModeType::DFCW))
-        {
-            restoration_owner = TestToneRestorationOwner::ManagedIdleNonWspr;
-        }
-    }
-
-    web_test_tone.store(true);
-
-    // Save previous mode so we can restore it later.
-    lastMode = config.mode;
-    test_tone_restoration_owner = restoration_owner;
-
-    try
-    {
-    wsprTransmitter.stopAndJoin();
-
-    ArgParserConfig selector_preparation_cfg;
-    WsprFrequencyEntry entry;
-    double dial_freq = 0.0;
-    double actual_rf_freq = 0.0;
-    if (has_explicit_frequency_source)
-    {
-        const TestToneFrequencyPlan &frequency_plan = *explicit_frequency_plan;
-        const TestToneSelectorPlan &selector_plan = *explicit_selector_plan;
-        dial_freq = frequency_plan.dial_frequency_hz.has_value()
-            ? static_cast<double>(*frequency_plan.dial_frequency_hz)
-            : static_cast<double>(frequency_plan.actual_rf_frequency_hz);
-        actual_rf_freq = static_cast<double>(frequency_plan.actual_rf_frequency_hz);
-        entry = WsprFrequencyEntry{};
-        entry.dial_frequency_hz = dial_freq;
-        entry.token = frequency_plan.band;
-        if (selector_plan.enabled)
-        {
-            entry.selector_gpio = selector_plan.config.gpio;
-            entry.selector_gpio_active_high = selector_plan.config.active_high;
-        }
-        selector_preparation_cfg.band_gpio = explicit_planning_snapshot->band_gpio;
-        selector_preparation_cfg.transmit_backend =
-            explicit_planning_snapshot->transmit_backend;
-        selector_preparation_cfg.allow_unqualified_frequency =
-            explicit_planning_snapshot->allow_unqualified_frequency;
-        selector_preparation_cfg.allow_non_amateur_frequency =
-            explicit_planning_snapshot->allow_non_amateur_frequency;
-        result.band = frequency_plan.band;
-        result.dial_frequency_hz = frequency_plan.dial_frequency_hz.value_or(0);
-        result.audio_offset_hz = frequency_plan.audio_offset_hz.value_or(0);
-        result.resolution_source = frequency_plan.resolution_source;
-        result.preset = frequency_plan.preset;
-    }
-    else
-    {
-        // Legacy requests retain their established scheduler behavior.
-        selector_preparation_cfg = config;
-        entry = next_frequency_entry(/*restart=*/true);
-        dial_freq = entry.dial_frequency_hz;
-        const double configured_actual_rf_freq = resolve_actual_rf_frequency_hz(
-            dial_freq,
-            selector_preparation_cfg.wspr.audio_offset_hz,
-            FrequencyPath::WsprDial);
-        actual_rf_freq = tone_request.frequency_hz.has_value()
-            ? static_cast<double>(*tone_request.frequency_hz)
-            : configured_actual_rf_freq;
-    }
-    current_frequency_entry = entry;
-
-    llog.logS(INFO, "Beginning test tone requested by web UI.");
-
-    // Switch into tone mode.
-    config.mode = ModeType::TONE;
-
-    llog.logS(
-        DEBUG,
-        "Resolved WSPR dial frequency ",
-        lookup.freq_display_string(dial_freq),
-        " to actual RF ",
-        lookup.freq_display_string(actual_rf_freq),
-        " using audio offset ",
-        explicit_frequency_plan.has_value()
-            ? static_cast<double>(explicit_frequency_plan->audio_offset_hz.value_or(0))
-            : selector_preparation_cfg.wspr.audio_offset_hz,
-        " Hz.");
-    if (tone_request.frequency_hz.has_value())
-    {
-        llog.logS(
-            INFO,
-            "Using web UI test tone RF frequency override: ",
-            lookup.freq_display_string(actual_rf_freq));
-    }
-    double committed_ppm = config.ppm;
-    if (transmit_backend_uses_gpio_output(config.transmit_backend))
-    {
-        const GpioFrequencyCorrection selected_correction =
-            select_and_publish_gpio_correction(config);
-        if (!selected_correction.valid)
-        {
-            throw std::runtime_error(selected_correction.reason);
-        }
-        committed_ppm = selected_correction.additional_ppm;
-        config.ppm = committed_ppm;
-    }
-    TransmissionRequest request =
-        make_tone_request(config, committed_ppm, actual_rf_freq, dial_freq, entry);
-    request.tone_duration = tone_request.duration;
-    if (explicit_frequency_plan.has_value())
-    {
-        request.applied_offset_hz = explicit_frequency_plan->audio_offset_hz.has_value()
-            ? static_cast<double>(*explicit_frequency_plan->audio_offset_hz)
-            : 0.0;
-    }
-    selector_preparation_cfg.mode = ModeType::TONE;
-    BandGPIOResolution selector_resolution;
-    const BandGPIOPrepareStatus selector_status =
-        prepare_band_gpio_for_frequency_or_log(
-            dial_freq,
-            entry,
-            selector_preparation_cfg,
-            -1,
-            &selector_resolution);
-    commit_band_gpio_snapshot_to_request(
-        request,
-        selector_resolution,
-        selector_status);
-    if (selector_status == BandGPIOPrepareStatus::Failed)
-    {
-        web_test_tone.store(false);
-        config.mode = lastMode;
-        test_tone_restoration_owner = TestToneRestorationOwner::Unknown;
-        const auto gpio_policy = wsprrypi::evaluate_gpio_band_policy(
-            to_controller_backend(selector_preparation_cfg.transmit_backend),
-            actual_rf_freq,
-            wsprrypi::TransmissionMode::TONE,
-            selector_preparation_cfg.allow_unqualified_frequency,
-            selector_preparation_cfg.allow_non_amateur_frequency,
-            to_controller_profile(selector_preparation_cfg.transmit_backend));
-        result.message = gpio_policy.allowed
-            ? "Unable to prepare the requested band selector."
-            : gpio_policy.error;
-        return result;
-    }
-    if (tone_request.rp1_development.enabled)
-    {
-        apply_test_tone_rp1_development_confirmation_bridge(
-            tone_request.rp1_development, config, request);
-    }
-    commit_execution_request(request);
-    result.actual_rf_frequency_hz = static_cast<std::uint64_t>(actual_rf_freq);
-    result.selector_gpio_enabled = request.selector_gpio_enabled;
-    result.selector_gpio = request.selector_gpio_enabled
-        ? request.selector_gpio_config.gpio
-        : -1;
-    result.selector_gpio_active_high = request.selector_gpio_enabled
-        ? request.selector_gpio_config.active_high
-        : false;
-    if (result.band.empty()) result.band = ham_band_to_string(*lookup.lookup_ham_band(dial_freq));
-
-    if (!suppress_scheduler_execution_for_test)
-    {
-        wsprTransmitter.startAsync();
-    }
-
-    llog.logS(INFO,
-              "WSPR-band test tone using dial frequency: ",
-              lookup.freq_display_string(dial_freq));
-    result.started = true;
-    result.message = "Test tone started.";
-    return result;
-    }
-    catch (const std::exception &error)
-    {
-        const std::string detail = error.what();
-        llog.logE(
-            ERROR,
-            "Test tone start rejected during configuration: " + detail);
-        rollback_failed_test_tone_start(lastMode);
-        result.message = detail.empty()
-            ? "Unable to configure the requested test tone."
-            : detail;
-        return result;
-    }
-    catch (...)
-    {
-        llog.logE(
-            ERROR,
-            "Test tone start rejected by an unknown configuration error.");
-        rollback_failed_test_tone_start(lastMode);
-        result.message = "Unable to configure the requested test tone.";
-        return result;
-    }
-}
-
-TestToneStartResult start_test_tone(
-    std::optional<std::uint64_t> frequency_hz_override)
-{
-    TestToneRequest request;
-    request.source = frequency_hz_override.has_value()
-        ? TestToneFrequencySource::LegacyExactRf
-        : TestToneFrequencySource::LegacyDefault;
-    request.frequency_hz = frequency_hz_override;
-    return start_test_tone(request);
-}
-
-/**
- * @brief End the transient runtime tone and restore prior orchestration.
- *
- * This stops the current tone, tears down selector lifecycle state through
- * the scheduler helper, and then resumes the pre-tone runtime mode.
- */
-TestToneStopResult end_test_tone()
-{
-    wsprrypi::invalidateRp1GpclkDevelopmentOperation();
-    TestToneStopResult result;
-    result.tone_was_active = web_test_tone.load();
-
-    if (!result.tone_was_active)
-    {
-        result.message = "No active test tone.";
-        return result;
-    }
-
-    llog.logS(INFO, "Ending test tone requested by Web UI.");
-
-    wsprTransmitter.stopAndJoin();
-    wsprTransmitter.clearExecutionStateAfterStop();
-    finalize_transmission_stop_cleanup(
-        &config,
-        runtime_should_hold_selector_gpios_initialized(config),
-        "test tone stop",
-        true,
-        true);
-    clear_committed_execution_request();
-    llog.logS(
-        DEBUG,
-        "Post-test-tone stop transmitter snapshot: ",
-        transmitter_reload_defer_debug_snapshot());
-    (void)reconcile_tx_led_after_transmitter_stop("test tone stop");
-
-    const bool deferred_reload_pending =
-        ini_reload_pending.load(std::memory_order_acquire);
-
-    const TestToneRestorationOwner restoration_owner = test_tone_restoration_owner;
-    test_tone_restoration_owner = TestToneRestorationOwner::Unknown;
-    web_test_tone.store(false);
-    config.mode = lastMode;
-
-    if (restoration_owner == TestToneRestorationOwner::WsprScheduler)
-    {
-        if (config.mode != ModeType::WSPR)
-        {
-            result.message = "Inconsistent Test Tone scheduler restoration state.";
-            return result;
-        }
-        if (!set_config(true))
-        {
-            result.message = "Unable to restore scheduler state after test tone stop.";
-            return result;
-        }
-
-        if (runtime_transmit_enabled(config) &&
-            !suppress_scheduler_execution_for_test)
-        {
-            // Re-arm the committed WSPR wait loop even if the tone stop
-            // interrupted a scheduler thread that had already been torn down.
-            wsprTransmitter.clearSoftOff();
-            wsprTransmitter.startAsync();
-        }
-
-        result.stopped = true;
-        result.scheduler_restored = true;
-        result.deferred_reload_reconciled =
-            deferred_reload_pending &&
-            !ini_reload_pending.load(std::memory_order_acquire);
-        result.message = "Test tone stopped and scheduler restored.";
-        return result;
-    }
-
-    if (restoration_owner == TestToneRestorationOwner::ManagedIdleNonWspr)
-    {
-        if (!(config.use_ini && !runtime_transmit_enabled(config) &&
-              (config.mode == ModeType::FSKCW || config.mode == ModeType::QRSS ||
-               config.mode == ModeType::DFCW)))
-        {
-            result.message = "Inconsistent managed Test Tone restoration state.";
-            return result;
-        }
-
-        result.stopped = true;
-        result.message = "Test tone stopped; managed mode restored inactive.";
-        return result;
-    }
-
-    if (restoration_owner != TestToneRestorationOwner::DirectToneStartup)
-    {
-        result.message = "Unknown Test Tone restoration state after safe cleanup.";
-        return result;
-    }
-
-    WsprFrequencyEntry entry;
-    double actual_rf_frequency_hz = 0.0;
-    if (!try_get_direct_tone_startup_request(
-            entry,
-            actual_rf_frequency_hz))
-    {
-        llog.logE(ERROR,
-                  "Unable to restore direct test tone; no "
-                  "transient tone request is active.");
-        result.message =
-            "Unable to restore direct test tone; no transient tone request is active.";
-        return result;
-    }
-
-    validate_config_data();
-    std::string restoration_error;
-    if (!start_direct_tone_execution(
-            config,
-            entry,
-            actual_rf_frequency_hz,
-            &restoration_error))
-    {
-        result.stopped = true;
-        result.message = restoration_error;
-        return result;
-    }
-
-    llog.logS(INFO,
-              "Transmitting tone, hit Ctrl-C to terminate tone.");
-    result.stopped = true;
-    result.message = "Test tone stopped and direct tone restored.";
-    return result;
-}
-
-StopTransmissionResult stop_transmission_by_user_request(bool persist_transmit)
-{
-    StopTransmissionResult result;
-    bool persist_to_ini = false;
-    suppress_cancelled_ws_event_for_user_stop.store(false, std::memory_order_release);
-
-    {
-        std::lock_guard<std::mutex> lk(set_config_mtx);
-
-        const WsprTransmitter::State state = wsprTransmitter.getState();
-        result.transmission_active =
-            state == WsprTransmitter::State::TRANSMITTING;
-
-        llog.logS(
-            INFO,
-            result.transmission_active
-                ? "Stop transmission requested by user; stopping active transmission."
-                : "Stop transmission requested by user; no active transmission.");
-
-        // Invalidate delayed launches before releasing the lock so no pending
-        // scheduler thread can start another transmission during stop handling.
-        non_wspr_schedule_generation.fetch_add(1, std::memory_order_acq_rel);
-        shutdown_after_current_transmission.store(false, std::memory_order_release);
-        shutdown_after_wspr_plan.store(false, std::memory_order_release);
-        reset_active_wspr_plan_state();
-
-        config.transmit = false;
-        result.transmit_disabled = true;
-        config_to_json();
-        persist_to_ini = config.use_ini && persist_transmit;
-    }
-
-    if (result.transmission_active)
-    {
-        suppress_cancelled_ws_event_for_user_stop.store(true, std::memory_order_release);
-    }
-
-    wsprTransmitter.stopAndJoin();
-    deassert_transmit_gpio_outputs(
-        &config,
-        false,
-        "scheduler shutdown");
-    release_idle_selector_gpio_reservations();
-
-    {
-        std::lock_guard<std::mutex> lk(set_config_mtx);
-
-        current_transmission_request = TransmissionRequest{};
-        current_dial_frequency = 0.0;
-        current_frequency_entry = WsprFrequencyEntry{};
-        freq_iterator = 0;
-        web_test_tone.store(false);
-
-        result.stop_performed = result.transmission_active;
-    }
-
-    if (persist_to_ini)
-    {
-        try
-        {
-            iniFile.set_bool_value("Operation", "Transmit", false);
-            iniFile.commit_changes();
-            result.persisted = true;
-            llog.logS(INFO, "Transmit disabled due to user stop request.");
-        }
-        catch (const std::exception &e)
-        {
-            result.persisted = false;
-            result.message =
-                std::string("Transmission stopped but failed to persist Operation.Transmit=false: ") +
-                e.what();
-            llog.logS(ERROR, result.message);
-            return result;
-        }
-    }
-    else
-    {
-        result.persisted = false;
-        result.message = persist_transmit
-                             ? "Transmission stopped and runtime transmit disabled; no INI file is active."
-                             : "Transmission stopped and runtime transmit disabled without persisting.";
-        llog.logS(INFO, result.message);
-        send_ws_message("transmit", "stopped");
-        return result;
-    }
-
-    {
-        std::lock_guard<std::mutex> lk(set_config_mtx);
-        set_managed_reload_tx_inhibited(false);
-    }
-    send_ws_message("transmit", "stopped");
-
-    result.message = result.transmission_active
-                         ? "Active transmission stopped and transmit disabled."
-                         : "Transmit disabled; no active transmission was running.";
-    return result;
-}
-
-static void stop_runtime_components_for_process_exit() noexcept
-{
-    ini_reload_pending.store(false, std::memory_order_relaxed);
-
-    webServer.stop();
-    socketServer.stop();
-    iniMonitor.stop();
-    shutdownMonitor.stop();
-    ppmManager.stop();
-    wsprTransmitter.shutdownForProcessExit();
-    deassert_transmit_gpio_outputs(
-        &config,
-        false,
-        "process-exit shutdown");
-    shutdown_all_configured_selector_gpios(config);
-    ampControl.stop();
-    ledControl.stop();
-}
-
-/**
- * @brief Main orchestration loop for startup, scheduling, and shutdown.
- *
- * @details
- * This loop validates configuration, starts long-lived services, prepares
- * the initial committed execution request, and then runs until shutdown.
- * WSPR startup goes through the same reload-safe scheduling path used for
- * later reconfiguration so request construction remains centralized here.
- *
- * @note This function blocks until `exitwspr_cv` is set by another thread.
- */
-bool wspr_loop()
-{
-    transmission_runtime_failed.store(false, std::memory_order_release);
-    const bool any_selector_gpio_configured =
-        has_configured_selector_gpios(config);
-
-    selector_gpio_control_enabled = any_selector_gpio_configured;
-    selector_gpio_drive_enabled = any_selector_gpio_configured;
-    bandGPIOSelector.setEnabled(selector_gpio_control_enabled);
-    bandGPIOSelector.setDriveGPIO(selector_gpio_drive_enabled);
-
-    // Display the final configuration after parsing arguments and INI file.
-    show_config_values();
-
-    const bool startup_config_handoff = consume_startup_config_handoff();
-    apply_managed_startup_policy_if_requested(startup_config_handoff);
-    set_startup_diagnostic_deferral(true);
-
-    if (config.mode != ModeType::WSPR)
-    {
-        if (startup_config_handoff)
-        {
-            apply_runtime_config_side_effects();
-        }
-        else
-        {
-            validate_config_data();
-        }
-    }
-    else
-    {
-        // Validate the startup WSPR configuration before any long-lived
-        // services are started so malformed CLI frequency lists fail cleanly.
-        if (startup_config_handoff)
-        {
-            apply_runtime_config_side_effects();
-        }
-        else
-        {
-            validate_config_data();
-        }
-    }
-
-    // Backend selection occurs while loading the validated runtime
-    // configuration.  Quiesce it before any service, scheduler, selector, or
-    // automatic-transmit path can run.  This latch is process-lifetime by
-    // design: ordinary reloads and toggles cannot clear a hardware-safety
-    // failure; a deliberate process restart reinitializes the backend.
-    if (!run_startup_quiesce_gate(config))
-    {
-        llog.logS(
-            ERROR,
-            "Startup transmission inhibition latched: ",
-            startup_quiesce_error);
-    }
-
-    if (config.transmit_backend == TransmitBackendKind::RP1_GPCLK)
-    {
-        const auto reconciliation = reconcile_rp1_idle_startup(config.gpio_tx_pin);
-        if (!reconciliation.ok)
-        {
-            llog.logS(ERROR,
-                "RP1 GPCLK startup reconciliation failed; transmission remains inhibited: ",
-                reconciliation.message);
-        }
-        else if (reconciliation.policy_domain == "startup-idle")
-        {
-            llog.logS(INFO,
-                "RP1 GPCLK route reconciled for safe idle startup; exact provider "
-                "identity and operation-scoped authorization remain required.");
-        }
-    }
-
-    const bool start_web = web_server_start_enabled(config);
-    const bool start_websocket = websocket_server_start_enabled(config);
-    const int startup_web_port = config.web_port;
-    const uint16_t startup_socket_port = config.socket_port;
-    const bool startup_socket_loopback_only = config.socket_loopback_only;
-    const auto startup_socket_loopback_family = config.socket_loopback_family;
-    if (!config.enable_web)
-    {
-        llog.logS(INFO, "Web UI disabled via CLI (--no-web)");
-    }
-    else if (start_web)
-    {
-        try
-        {
-            webServer.start(startup_web_port);
-            webServer.setThreadPriority(SCHED_RR, 10);
-        }
-        catch (const std::exception &error)
-        {
-            llog.logE(ERROR, "Unable to start HTTP listener: ", error.what());
-            webServer.stop();
-            stop_runtime_components_for_process_exit();
-            return false;
-        }
-        const auto listener_deadline = std::chrono::steady_clock::now() +
-                                       std::chrono::seconds(5);
-        while (!webServer.isListening() &&
-               !exiting_wspr.load(std::memory_order_acquire) &&
-               std::chrono::steady_clock::now() < listener_deadline)
-        {
-            std::this_thread::sleep_for(std::chrono::milliseconds(20));
-        }
-        if (!webServer.isListening())
-        {
-            llog.logE(ERROR, "HTTP listener failed to bind or listen.");
-            webServer.stop();
-            stop_runtime_components_for_process_exit();
-            return false;
-        }
-        llog.logS(INFO, "HTTP listener started on port ", startup_web_port, ".");
-    }
-    else
-    {
-        llog.logS(DEBUG, "Skipping web server.");
-    }
-
-    if (start_websocket)
-    {
-        if (!socketServer.start(
-                startup_socket_port,
-                SOCKET_KEEPALIVE,
-                startup_socket_loopback_only,
-                startup_socket_loopback_family))
-        {
-            llog.logE(ERROR, "WebSocket listener failed to start.");
-            stop_runtime_components_for_process_exit();
-            return false;
-        }
-        socketServer.setThreadPriority(SCHED_RR, 10);
-    }
-    else if (!start_websocket)
-    {
-        llog.logS(DEBUG, "Skipping socket server.");
-    }
-
-    if ((start_web || start_websocket) &&
-        active_privileged_network_mode() ==
-        PrivilegedNetworkMode::insecure_disabled)
-    {
-        llog.logS(WARN, "NETWORK SAFETY OFF");
-    }
-    else if (start_web || start_websocket)
-    {
-        const auto network_snapshot =
-            SupportRequestGuard::discover_local_networks();
-        if (!network_snapshot.discovery_succeeded)
-        {
-            llog.logS(WARN,
-                "Privileged network safety is enforced; current interface "
-                "discovery failed. Listeners remain available and protected "
-                "non-loopback clients are rejected.");
-        }
-        else if (network_snapshot.networks.empty())
-        {
-            llog.logS(INFO,
-                "Privileged network safety is enforced with no eligible LAN. "
-                "Listeners remain available for loopback; protected "
-                "non-loopback clients are rejected.");
-        }
-    }
-
-    // Set transmission server and set priority
-    wsprTransmitter.setThreadScheduling(SCHED_FIFO, 50);
-
-    // Set transmission event callbacks
-    wsprTransmitter.setTransmissionCallbacks(
-        [](WsprTransmitter::TransmissionCallbackEvent event,
-           WsprTransmitter::LogLevel level,
-           const std::string &msg,
-           double value)
-        {
-            transmitter_cb(event, level, msg, value);
-        });
-
-    // Monitor INI file for changes
-    if (config.use_ini)
-    {
-        // Start INI monitor
-        iniMonitor.filemon(config.ini_filename, callback_ini_changed);
-        iniMonitor.setPriority(SCHED_RR, 10);
-    }
-
-    llog.logS(INFO, "WSPR loop running.");
-    set_startup_diagnostic_deferral(false);
-    emit_deferred_startup_diagnostics();
-
-    // Startup WSPR configuration should be applied exactly once using the
-    // same reload-safe path that handles validation, setup, and scheduling.
-    if (config.mode == ModeType::WSPR)
-    {
-        ini_reload_pending.store(!startup_config_handoff, std::memory_order_relaxed);
-        if (!set_config(startup_config_handoff ? false : true))
-        {
-            stop_runtime_components_for_process_exit();
-            return false;
-        }
-    }
-    else if (config.mode == ModeType::TONE)
-    {
-        log_scheduler_path_selection(config.mode);
-        WsprFrequencyEntry entry;
-        double actual_rf_frequency_hz = 0.0;
-        if (!try_get_direct_tone_startup_request(entry, actual_rf_frequency_hz))
-        {
-            llog.logE(ERROR, "Direct RF test tone requested without a startup tone request.");
-            stop_runtime_components_for_process_exit();
-            return false;
-        }
-
-        std::string startup_error;
-        if (!start_direct_tone_execution(
-                config,
-                entry,
-                actual_rf_frequency_hz,
-                &startup_error))
-        {
-            llog.logS(ERROR, startup_error);
-            stop_runtime_components_for_process_exit();
-            return false;
-        }
-        llog.logS(INFO, "transmitting tone, hit Ctrl-C to terminate tone.");
-    }
-    else if (config.mode == ModeType::QRSS)
-    {
-        log_scheduler_path_selection(config.mode);
-        if (has_non_wspr_cli_startup_request(config.mode))
-        {
-            shutdown_after_current_transmission.store(true, std::memory_order_release);
-            shutdown_after_wspr_plan.store(false, std::memory_order_release);
-            if (!start_non_wspr_transmission_now(config))
-            {
-                stop_runtime_components_for_process_exit();
-                return false;
-            }
-        }
-        else
-        {
-            shutdown_after_current_transmission.store(false, std::memory_order_release);
-            shutdown_after_wspr_plan.store(false, std::memory_order_release);
-            schedule_next_non_wspr_launch(config);
-        }
-    }
-    else if (config.mode == ModeType::FSKCW)
-    {
-        log_scheduler_path_selection(config.mode);
-        if (has_non_wspr_cli_startup_request(config.mode))
-        {
-            shutdown_after_current_transmission.store(true, std::memory_order_release);
-            shutdown_after_wspr_plan.store(false, std::memory_order_release);
-            if (!start_non_wspr_transmission_now(config))
-            {
-                stop_runtime_components_for_process_exit();
-                return false;
-            }
-        }
-        else
-        {
-            shutdown_after_current_transmission.store(false, std::memory_order_release);
-            shutdown_after_wspr_plan.store(false, std::memory_order_release);
-            schedule_next_non_wspr_launch(config);
-        }
-    }
-    else if (config.mode == ModeType::DFCW)
-    {
-        log_scheduler_path_selection(config.mode);
-        if (has_non_wspr_cli_startup_request(config.mode))
-        {
-            shutdown_after_current_transmission.store(true, std::memory_order_release);
-            shutdown_after_wspr_plan.store(false, std::memory_order_release);
-            if (!start_non_wspr_transmission_now(config))
-            {
-                stop_runtime_components_for_process_exit();
-                return false;
-            }
-        }
-        else
-        {
-            shutdown_after_current_transmission.store(false, std::memory_order_release);
-            shutdown_after_wspr_plan.store(false, std::memory_order_release);
-            schedule_next_non_wspr_launch(config);
-        }
-    }
-
-    if (const char *restoration = std::getenv("WSPRRYPI_ROUTE_RESTORE_IDLE"))
-    {
-        const auto listener_deadline = std::chrono::steady_clock::now() + std::chrono::seconds(5);
-        while (start_web && !webServer.isListening() &&
-               std::chrono::steady_clock::now() < listener_deadline)
-            std::this_thread::sleep_for(std::chrono::milliseconds(20));
-        if (startup_quiesce_inhibited.load(std::memory_order_acquire) ||
-            (start_web && !webServer.isListening()) ||
-            config.transmit_backend != TransmitBackendKind::RP1_GPCLK ||
-            !acknowledge_rp1_restoration(restoration, config.transmit))
-        {
-            llog.logS(ERROR, "Route restoration could not acknowledge idle application readiness.");
-            stop_runtime_components_for_process_exit();
-            return false;
-        }
-    }
-
-    // -------------------------------------------------------------------------
-    // Loop (block wspr_loop only) until shutdown is triggered
-    // -------------------------------------------------------------------------
-    {
-        std::unique_lock<std::mutex> lk(exitwspr_mtx);
-        exitwspr_cv.wait(lk, []
-                         { return exitwspr_ready; });
-    }
-
-    llog.logS(DEBUG, "WSPR loop termination started.");
-
-    // -------------------------------------------------------------------------
-    // Shutdown and cleanup
-    // -------------------------------------------------------------------------
-    llog.logS(DEBUG, "Stopping runtime components.");
-
-    llog.logS(DEBUG, "Stopping web server.");
-    webServer.stop();
-    llog.logS(DEBUG, "Web server stopped.");
-
-    llog.logS(DEBUG, "Stopping socket server.");
-    socketServer.stop();
-    llog.logS(DEBUG, "Socket server stopped.");
-
-    llog.logS(DEBUG, "Stopping configuration monitor.");
-    ini_reload_pending.store(false, std::memory_order_relaxed);
-    iniMonitor.stop(); // Stop config file monitor before transmitter teardown.
-    llog.logS(DEBUG, "Configuration monitor stopped.");
-
-    llog.logS(DEBUG, "Stopping shutdown monitor.");
-    shutdownMonitor.stop(); // Stop the GPIO monitor
-    llog.logS(DEBUG, "Shutdown monitor stopped.");
-
-    llog.logS(DEBUG, "Stopping PPM manager.");
-    ppmManager.stop(); // Stop PPM manager (if active)
-    llog.logS(DEBUG, "PPM manager stopped.");
-
-    llog.logS(DEBUG, "Stopping transmitter.");
-    wsprTransmitter.shutdownForProcessExit();
-    deassert_transmit_gpio_outputs(
-        &config,
-        false,
-        "scheduler process shutdown");
-    llog.logS(DEBUG, "Transmitter stopped.");
-
-    llog.logS(DEBUG, "Stopping band GPIO selector.");
-    shutdown_all_configured_selector_gpios(config);
-    llog.logS(DEBUG, "Band GPIO selector stopped.");
-
-    llog.logS(DEBUG, "Stopping Amp Control driver.");
-    ampControl.stop();
-    llog.logS(DEBUG, "Amp Control driver stopped.");
-
-    llog.logS(DEBUG, "Stopping LED driver.");
-    ledControl.stop(); // Stop LED driver
-    llog.logS(DEBUG, "LED driver stopped.");
-
-    llog.logS(DEBUG, "Runtime components stopped.");
-
-    llog.logS(INFO, get_project_name(), "exiting.");
-    // Flush all file system buffers to disk
-    sync();
-
-    return !transmission_runtime_failed.load(std::memory_order_acquire);
-}
-
-/**
- * @brief Synchronize disk and reboot the machine.
- *
- * This function calls sync() to flush filesystem buffers, then
- * invokes the reboot(2) syscall directly. The process must have
- * the CAP_SYS_BOOT capability (typically run as root).
- */
-void reboot_machine()
-{
-    const MachinePowerResult result = request_machine_power(MachinePowerOperation::Reboot);
-    if (result.status == MachinePowerStatus::Unsupported)
-    {
-        llog.logE(ERROR,
-                  "Machine reboot is unavailable on this platform; "
-                  "application shutdown completed without rebooting the machine.");
-    }
-    else if (result.status == MachinePowerStatus::Failed)
-    {
-        llog.logE(ERROR, "Reboot failed: ", std::strerror(result.error_number));
-    }
-}
-
-/**
- * @brief Flush filesystems and power off the machine.
- *
- * Calls sync() to ensure all disk buffers are written, then invokes
- * the reboot(2) syscall with the POWER_OFF command. Requires root or
- * the CAP_SYS_BOOT capability.
- */
-void shutdown_machine()
-{
-    const MachinePowerResult result = request_machine_power(MachinePowerOperation::PowerOff);
-    if (result.status == MachinePowerStatus::Unsupported)
-    {
-        llog.logE(ERROR,
-                  "Machine power-off is unavailable on this platform; "
-                  "application shutdown completed without powering off the machine.");
-    }
-    else if (result.status == MachinePowerStatus::Failed)
-    {
-        llog.logE(ERROR, "Shutdown failed: ", std::strerror(result.error_number));
-    }
-}
-
-
 static std::string runtime_mode_to_string(
     wsprrypi::TransmissionMode mode)
 {
@@ -4846,8 +3305,8 @@ WsprRuntimeStatusSnapshot current_tx_runtime_status_snapshot()
         snapshot.gpio_correction_candidate = current_gpio_candidate_provenance;
         snapshot.gpio_correction_committed = committed_gpio_correction_provenance;
     }
-    snapshot.tx_state = wsprTransmitter.stateToStringLower(
-        wsprTransmitter.getState());
+    snapshot.tx_state = transmitter_state_string_lower(
+        transmitter_state());
     snapshot.gpio_correction_committed.active =
         snapshot.gpio_correction_committed.available &&
         snapshot.tx_state == "transmitting";
@@ -4856,7 +3315,7 @@ WsprRuntimeStatusSnapshot current_tx_runtime_status_snapshot()
     {
         snapshot.gpio_correction_committed = {};
     }
-    const auto runtime_status = wsprTransmitter.runtimeExecutionStatusSnapshot();
+    const auto runtime_status = transmitter_runtime_status();
     if (snapshot.tx_state == "transmitting")
     {
         snapshot.runtime_mode = runtime_mode_to_string(runtime_status.mode);
@@ -5034,11 +3493,12 @@ bool rp1_route_transaction_inhibited_state() noexcept
 wsprrypi::Rp1GpclkApplicationIdleState rp1_gpclk_application_idle_state() noexcept
 {
     wsprrypi::Rp1GpclkApplicationIdleState state;
-    const auto transmitter_state = wsprTransmitter.getState();
+    const auto current_transmitter_state = transmitter_state();
     state.controller_prepared = config.transmit;
-    const bool controller_quiescent = transmitter_state == WsprTransmitState::DISABLED ||
-        transmitter_state == WsprTransmitState::COMPLETE ||
-        transmitter_state == WsprTransmitState::CANCELLED;
+    const bool controller_quiescent =
+        current_transmitter_state == WsprTransmitState::DISABLED ||
+        current_transmitter_state == WsprTransmitState::COMPLETE ||
+        current_transmitter_state == WsprTransmitState::CANCELLED;
     state.execution_active = !controller_quiescent || web_test_tone.load(std::memory_order_acquire);
     state.schedule_committed = config.transmit || !active_wspr_plan.frames.empty();
     state.stop_or_drain_active = exiting_wspr.load(std::memory_order_acquire);
@@ -5210,23 +3670,6 @@ bool selector_gpio_logical_state_for_test(
 void stop_active_transmission_selectors_for_test() noexcept
 {
     stop_active_transmission_selectors();
-}
-
-void stop_runtime_components_for_test() noexcept
-{
-    webServer.stop();
-    socketServer.stop();
-    iniMonitor.stop();
-    shutdownMonitor.stop();
-    ppmManager.stop();
-    wsprTransmitter.stopAndJoin();
-    deassert_transmit_gpio_outputs(
-        &config,
-        false,
-        "test runtime stop");
-    ampControl.stop();
-    ledControl.stop();
-    release_idle_selector_gpio_reservations();
 }
 
 bool park_active_transmission_selectors_for_test() noexcept
@@ -5430,10 +3873,20 @@ bool run_startup_quiesce_gate_for_test(const ArgParserConfig &cfg)
 
 bool startup_quiesce_inhibited_for_test() noexcept
 {
-    return startup_quiesce_inhibited.load(std::memory_order_acquire);
+    return startup_quiesce_inhibited_state();
 }
 
 std::string startup_quiesce_error_for_test()
+{
+    return startup_quiesce_error_state();
+}
+
+bool startup_quiesce_inhibited_state() noexcept
+{
+    return startup_quiesce_inhibited.load(std::memory_order_acquire);
+}
+
+std::string startup_quiesce_error_state()
 {
     std::lock_guard<std::mutex> lock(startup_quiesce_error_mtx);
     return startup_quiesce_error;
